Test renamed to VaadinSmokeTest
